backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.8.0
eventlet==0.35.2
librosa==0.10.1
numpy==1.26.0
scipy==1.11.4
soundfile==0.12.1
openai-whisper>=20250625
torch>=2.7.0
torchaudio==2.1.0
pydub==0.25.1
python-multipart==0.0.6
python-dotenv==1.0.0
werkzeug==3.0.0
matplotlib>=3.7.0
